Create a 'FeatureNotifications' component to use in Notifications component
